third_party/WebKit/LayoutTests/batterystatus/multiple-frames-expected.txt

CONSOLE MESSAGE: line 47: Uncaught TypeError: Cannot call method 'addEventListener' of undefined
Tests using BatteryStatus from multiple frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
